feat: 数据库迁移与模型层完善
- 扩展建表迁移，新增字典/字典详情/操作记录/参数/错误日志等表
- 新增 Casbin API 种子数据迁移（m20240101_000003_seed_api_casbin.rs）
- 修复种子数据与 gin-vue-admin 对齐（菜单排序/图标/keep_alive 等）
- 新增 Model：sys_authority_btn/sys_base_menu_btn/sys_base_menu_parameter
- 新增 Model：sys_data_authority_id/sys_dictionary/sys_dictionary_detail
- 新增 Model：sys_error/sys_ignore_api/sys_operation_record/sys_params
- 新增 Model：sys_user_authority
- 所有 Model 统一添加 serde rename_all camelCase 和 ID 大写序列化

+++ server/src/model/system/casbin_rule.rs
@@ -1,14 +1,15 @@
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
-/// Casbin 规则表（casbin_rules）
+/// Casbin 规则表（casbin_rule）
 /// 对应 Gin-Vue-Admin 的 CasbinRule
 /// 格式：ptype="p", v0=角色ID, v1=路径, v2=方法
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
-#[sea_orm(table_name = "casbin_rules")]
+#[sea_orm(table_name = "casbin_rule")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    #[serde(rename = "ID")]
+    pub id: u64,
     /// 策略类型：p（权限策略）或 g（角色继承）
     pub ptype: String,
     pub v0: String,
