+++ server/src/model/system/sys_api.rs
@@ -7,11 +7,15 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(table_name = "sys_apis")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    #[serde(rename = "ID")]
+    pub id: u64,
+    #[serde(rename = "CreatedAt")]
     pub created_at: Option<DateTimeUtc>,
+    #[serde(rename = "UpdatedAt")]
     pub updated_at: Option<DateTimeUtc>,
+    #[serde(skip_serializing)]
     pub deleted_at: Option<DateTimeUtc>,
-    /// API 路径，如 /user/getUserInfo
+    /// API 路径    /// API 路径，如 /user/getUserInfo
     pub path: String,
     /// API 描述
     pub description: String,
